re-introduce optimistic mode
the switch is now optimistic, but use_current is still in use.
Once use_current fails to verify, the switch is returned to the old position.

Additionally, the switch now has an extra attribute (pending_modechange)
that contains the new mode as long as the state is optimistic.
If the state is not optimistic, the attribute contians "None"